[BUGFIX] htmlArea RTE: Relative url's on pasted links/images in Firefox
When copying/pasting an internal link or image from and to the RTE
editing area, Firefox changes the href or src to a relative url.
After save, the url ends up being wrong.

Similar issues arises when moving images or links inside the editing
area and when dropping images in the RTE from the image dialogue
window.

This change works around Firefox issue
https://bugzilla.mozilla.org/show_bug.cgi?id=613517,
by making the url absolute again before saving or toggling
to source text mode, as well as on paste and drop events.

Change-Id: I27024687d897a36133113f47b4d52ca9e0ab6c90
Resolves: #28497, #24852, #27353
Releases: 4.4, 4.5, 4.6
Reviewed-on: http://review.typo3.org/3916
Reviewed-by: Stanislas Rolland
Tested-by: Stanislas Rolland