[FEATURE] Add hooks to extend browse_links.php
Add hooks to be able to extend the browse links popup in RTE and for
TCEForms link fields.

This is useful to add new fields to the link wizard to add additional
params, onclick attributes, rel attributes or other tracking parameters
to the rendered link.

This requires hooks also in rtehtmlarea.

Change-Id: I917829cc3be423b4ae2e92a3264894152f912fcd
Resolves: #32809
Releases: 4.7
Reviewed-on: http://review.typo3.org/7495
Reviewed-by: Stefan Neufeind
Reviewed-by: Wouter Wolters
Tested-by: Martin Tepper
Reviewed-by: Joh. Feustel
Tested-by: Joh. Feustel
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch