[CLEANUP] Remove deprecated filemount functions in BE_USER
With the introduction of the file abstraction layer,
the filemount functionality was deprecated in 6.0
but wasn't removed.

However, the usage of addFileMount() in  rtehtmlarea and
EXT:recordlist is now removed, since the added filemounts
are never read anywhere, so this is only dead code.

The functionality needs to be re-implemented.

Releases: 6.2
Resolves: #56772
Change-Id: I20ec66f46f15a785f5014cb625dabef711fae0a1
Reviewed-on: https://review.typo3.org/28265
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Oliver Klee
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring