[TASK] RTE: Use CSS3 resize when possible
This works fine in Firefox. In all other browsers, we need to fallback
to jQuery UI Resizable, because IE 11 does not support this property,
and in WebKit and Opera there are bugs in this special case in
Mutation Observer (required to resize the iframe which does not
follow automatically).

Releases: master
Resolves: #64470
Change-Id: Iac1f4d721908967f2bc10429c6a3b716eb33adbe
Reviewed-on: http://review.typo3.org/36225
Reviewed-by: Stanislas Rolland <typo3@sjbr.ca>
Tested-by: Stanislas Rolland <typo3@sjbr.ca>