[TASK] Remove obsolete class aliases
The class aliases introduced during RTE and form engine refactoring
are more or less useless as the API changed heavily in this area
so that a class alias alone would provide no benefit.

These changes were marked as breaking already, so we can do this
now as a regular cleanup.

Releases: master
Resolves: #67967
Change-Id: Ied68f390aa9a33be3a8bb9b26732efb698e04848
Reviewed-on: http://review.typo3.org/40976
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>