[FEATURE] Add Dispatcher for Backend Routing
The newly introduced Backend Routing is updated so that Routes
must define not just the controller but also the action which should
be called. It is now also possible to use a Closure / callable directly
to be called as a Route.

The RequestHandler now creates the Response object, which is
handed to the Dispatcher and to the respective action. This way
each action can already use the existing Response object.

The Dispatcher resolves the action and hands over the request and
response object.

All existing routes have been adapted to use controller and actions.

The fixed ControllerInterface is not needed anymore.

Resolves: #69855
Releases: master
Change-Id: I6edcf8e0f6b49580052f24ca9fd7ad352b6d6f34
Reviewed-on: http://review.typo3.org/43337
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>