[!!!][TASK] RTE transformations: Remove hsc options
The option to re-convert htmlspecialchars() between transformations
RTE and DB is always on, in order to clean HTML is always turned
on, meaning that the HTMLcleaner() method will not do htmlspecialchars() at all by default.

As these additional options is now obsolete with the current rtehtmlarea
extension (thus, always "enabled"), the options are removed.

Resolves: #72889
Releases: master
Change-Id: I6092054c65a031615546ceff9ee70a94f3033ec7
Reviewed-on: https://review.typo3.org/46145
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>