[BUGFIX] RTE: Correct behavior on empty textfield in Firefox
With Firefox, the RTE produces several JavaScript errors if the textfield
is empty. The issue is that on Firefox it's tried to access the parent
node of the html tag which is nonsense.

Some regular expressions and an additional 'undefined' check are added
to fix this wrong behavior.

Resolves: #77191
Releases: master, 7.6
Change-Id: I617bdb6415c0468894aee08101c9d26cb790a077
Reviewed-on: https://review.typo3.org/49150
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Bamboo TYPO3com <info@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>