[!!!][TASK] Remove RteHtmlParser proc options
The RTE processing options "dontProtectUnknownTags_rte"
and "dontConvAmpInNBSP_rte" come from a legacy point of
view when transforming from DB to the RTE and have never
been set since TYPO3 4.x as a default for installations.

Additionally the option "dontConvBRtoParagraph" which was used
for setups without RTE where people just wrote plain HTML is
removed, so BR tags are kept as they are (which is a regular use case
in any RTE). This option was set with HtmlArea at all time
(same with other RTEs).

The options are very custom and when transferring data to the RTE
the best practice is applied by default since a decade (original code
was from CVS by Kasper - before the RTE was enabled by default).

Resolves: #79273
Releases: master
Change-Id: Ib05955b7294fc8f5600d91de443f519176b58634
Reviewed-on: https://review.typo3.org/51264
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>